libft/lambdas/finance/sutton-pgp/requirements.txt
boto3
requests
jsonpickle
aws-xray-sdk
cachetools
python-gnupg
